Revert "Feature singer"
