+++ node-backend/index.js
@@ -1,21 +1,11 @@
 const registerRecommend = require('./recommend')
-const {
-  registerSinger,
-  registerSingerDetail
-} = require('./singer')
-
-const {
-  registerSongsUrl
-} = require('./song')
+const registerSinger = require('./singer')
 
 // 注册后端路由
 function registerRouter (app) {
   registerRecommend(app)
 
   registerSinger(app)
-  registerSingerDetail(app)
-
-  registerSongsUrl(app)
 }
 
 module.exports = registerRouter
