+++ node-backend/singer/index.js
@@ -7,8 +7,6 @@ const { get } = require('../request')
 const { getRandomVal } = require('../utils')
 
 const pinyin = require('pinyin')
-// 歌曲图片加载失败时使用的默认图片
-const fallbackPicUrl = 'https://y.gtimg.cn/mediastyle/music_v11/extra/default_300x300.jpg?max_age=31536000'
 
 // 响应成功code
 const CODE_OK = 0
@@ -150,86 +148,6 @@ function registerSinger(app) {
   })
 }
 
-// 注册歌手详情接口路由
-function registerSingerDetail (app) {
-  app.get('/api/getSingerDetail', (req, res) => {
-    console.log('getSingerDetail ',req.query.mid)
-    const url = 'https://u.y.qq.com/cgi-bin/musics.fcg'
-
-    const data = JSON.stringify({
-      comm: {
-        ct: 24,
-        cv: 0
-      },
-      singerSongList: {
-        method: 'GetSingerSongList',
-        param: {
-          order: 1,
-          singerMid: req.query.mid,
-          begin: 0,
-          num: 100
-        },
-        module: 'musichall.song_list_server'
-      }
-    })
-
-    const randomKey = getRandomVal('getSingerSong')
-    const sign = getSecuritySign(data)
-
-    get(url, {
-      sign,
-      '-': randomKey,
-      data
-    }).then((response) => {
-      const data = response.data
-      if (data.code === CODE_OK) {
-        const list = data.singerSongList.data.songList
-        // 歌单详情、榜单详情接口都有类似处理逻辑，固封装成函数
-        const songList = handleSongList(list)
-
-        res.json({
-          code: CODE_OK,
-          result: {
-            songs: songList
-          }
-        })
-      } else {
-        res.json(data)
-      }
-    })
-  })
-}
-
-// 处理歌曲列表
-function handleSongList (list) {
-  const songList = []
-
-  list.forEach((item) => {
-    const info = item.songInfo || item
-    if (info.pay.pay_play !== 0 || !info.interval) {
-      // 过滤付费歌曲和获取不到时长的歌曲
-      return
-    }
-
-    // 构造歌曲的数据结构
-    const song = {
-      id: info.id,
-      mid: info.mid,
-      name: info.name,
-      singer: mergeSinger(info.singer),
-      url: '', // 在另一个接口获取
-      duration: info.interval,
-      pic: info.album.mid ? `https://y.gtimg.cn/music/photo_new/T002R800x800M000${info.album.mid}.jpg?max_age=2592000`
-        : fallbackPicUrl,
-      album: info.album.name
-    }
-
-    songList.push(song)
-  })
-
-  return songList
-}
-
 // 格式化歌手列表
 function normalizeSingers(singers) {
   if (Array.isArray(singers)) {
@@ -240,32 +158,16 @@ function normalizeSingers(singers) {
   return createSinger(singers)
 }
 
-// 合并多个歌手的姓名
-function mergeSinger (singer) {
-  const ret = []
-  if (!singer) {
-    return ''
-  }
-  singer.forEach((s) => {
-    ret.push(s.name)
-  })
-  return ret.join('/')
-}
-
 function createSinger(singer) {
   return new Singer(singer)
 }
-
 class Singer {
   constructor(singer) {
     this.id = singer.singer_id
     this.mid = singer.singer_mid
     this.name = singer.singer_name
-    this.pic = singer.singer_pic.replace(/\.webp$/, '.jpg').replace('150x150', '800x800')
+    this.pic = singer.singer_pic
   }
 }
 
-module.exports = {
-  registerSinger,
-  registerSingerDetail
-}
+module.exports = registerSinger
