+++ package.json
@@ -14,7 +14,6 @@
     "axios": "^0.21.1",
     "core-js": "^3.6.5",
     "fibers": "^5.0.0",
-    "good-storage": "^1.1.1",
     "postcss": "^8.2.15",
     "sass": "^1.32.12",
     "style-loader": "^2.0.0",
