--- src/assets/js/constant.js
@@ -1,1 +1,0 @@
-export const SINGER_KEY = '__singer__'
