+++ src/components/base/loading/loading.vue
@@ -22,12 +22,12 @@ export default {
     const loadingMsg = ref('正在载入中...')
 
     // 对loading directive提供的内部方法, 来动态修改loadingMessage的值
-    const setTitle = msg => {
+    const setLoadingMsg = msg => {
       loadingMsg.value = msg
     }
     return {
       loadingMsg,
-      setTitle
+      setLoadingMsg
     }
   }
 }
