--- src/components/base/no-result/directive.js
@@ -1,6 +1,0 @@
-import NoResult from './no-result'
-import createLoadingLikeDirective from '@/assets/js/create-loading-like-directive'
-
-const noResultDirective = createLoadingLikeDirective(NoResult)
-
-export default noResultDirective
