+++ src/components/singer-list/singer-list.vue
@@ -17,7 +17,6 @@
             v-for="item in group.list"
             :key="item.id"
             class="list__group__item"
-            @click="onItemClick(item)"
           >
             <img class="avatar" :src="item.pic" alt="">
             <span class="name">{{ item.name }}</span>
@@ -67,7 +66,6 @@ export default {
   components: {
     Scroller
   },
-  emits: ['clickSinger'],
   props: {
     data: {
       type: Array,
@@ -92,12 +90,6 @@ export default {
       onShortcutTouchEnd
     } = useShortcut(props, groupRef)
 
-    const onItemClick = (item) => {
-      console.log('onItemClick')
-      console.log(item)
-      emit('clickSinger', item)
-    }
-
     return {
       groupRef,
       scrollRef,
@@ -109,8 +101,7 @@ export default {
       onScroll,
       onShortcutTouchStart,
       onShortcutTouchMove,
-      onShortcutTouchEnd,
-      onItemClick
+      onShortcutTouchEnd
     }
   }
 }
@@ -194,7 +185,7 @@ $shortcutColor: var(--shortcutColor, #ffcd32);
       }
     }
 
-    &__hint {
+    &__hint{
       position: absolute;
       top: 50%;
       right: 200%;
