+++ src/router/index.js
@@ -3,7 +3,6 @@ import Recommend from '@/views/recommend'
 import Singer from '@/views/singer'
 import TopList from '@/views/top-list'
 import Search from '@/views/search'
-import SingerDetail from '@/views/singer-detail'
 
 const routes = [
   {
@@ -16,13 +15,7 @@ const routes = [
   },
   {
     path: '/singer',
-    component: Singer,
-    children: [
-      {
-        path: ':id',
-        component: SingerDetail
-      }
-    ]
+    component: Singer
   },
   {
     path: '/top-list',
