--- src/service/song.js
@@ -1,23 +1,0 @@
-// 歌曲相关
-import { get } from './base'
-
-export function processSongs (songs) {
-  if (!songs.length) {
-    return Promise.resolve(songs)
-  }
-
-  return get('/api/getSongsUrl', {
-    mid: songs.map((song) => {
-      return song.mid
-    })
-  }).then((result) => {
-    const map = result.map
-    return songs.map((song) => {
-      song.url = map[song.mid]
-      return song
-    }).filter((song) => {
-      // 存在该字段才可以播放
-      return song.url.indexOf('vkey') > -1
-    })
-  })
-}
